+++ src/Components/Facility/FacilityHome.tsx
@@ -5,7 +5,6 @@ import loadable from "@loadable/component";
 import Dialog from "@material-ui/core/Dialog";
 import DialogActions from "@material-ui/core/DialogActions";
 import DialogTitle from "@material-ui/core/DialogTitle";
-import RadialCard from "../Common/components/RadialCard";
 import DialogContent from "@material-ui/core/DialogContent";
 import DialogContentText from "@material-ui/core/DialogContentText";
 import {
@@ -141,7 +140,6 @@ export const FacilityHome = (props: any) => {
     return <Loading />;
   }
   let capacityList: any = null;
-  let percentage = 0.0;
   let totalBedCount = 0;
   let totalOccupiedBedCount = 0;
   if (!capacityData || !capacityData.length) {
@@ -155,11 +153,10 @@ export const FacilityHome = (props: any) => {
       totalBedCount += x.total_capacity ? x.total_capacity : 0;
       totalOccupiedBedCount += x.current_capacity ? x.current_capacity : 0;
     });
-    // percentage = (totalOccupiedBedCount * 100) / totalBedCount;
 
     capacityList = (
       <div className="mt-4 grid lg:grid-cols-3 sm:grid-cols-2 gap-7 w-full">
-        <RadialCard
+        <BedTypeCard
           label={"Total Beds"}
           bedCapacityId={0}
           used={totalOccupiedBedCount}
@@ -178,7 +175,7 @@ export const FacilityHome = (props: any) => {
             console.log(res);
             return (
               <div>
-                <RadialCard
+                <BedTypeCard
                   facilityId={facilityId}
                   bedCapacityId={res.id}
                   key={`bed_${res.id}`}
@@ -634,33 +631,6 @@ export const FacilityHome = (props: any) => {
             Add More Bed Types
           </ButtonV2>
         </div>
-        {totalBedCount !== 0 && (
-          <div className="mb-8">
-            {/* <RadialCard
-        // isLoading={isLoading}
-        label={"Total Beds"}
-        used={totalOccupiedBedCount}
-        total={totalBedCount}
-        // previous={"used": totalOccupiedBedCount, "total": totalBedCount}
-        reverseIndicator
-      /> */}
-            {/* <div className="flex gap-1 text-md text-gray-600 font-semibold">
-        <p>Total Beds: </p>
-        <p className="text-black">
-          {totalOccupiedBedCount} / {totalBedCount}
-        </p>
-      </div>
-      <p className="text-xs">Total Occupied Beds / Total Beds</p>
-      <div className="rounded-full w-full bg-[#EDEEF0] mt-2">
-        <div
-          className="p-[6px] rounded-full bg-primary transition-all duration-500 ease-out"
-          style={{
-            width: `${percentage}%`,
-          }}
-        ></div>
-      </div> */}
-          </div>
-        )}
         <div>{capacityList}</div>
       </div>
       <div className="bg-white rounded p-3 md:p-6 shadow-sm mt-5">
